add topChecked.js
